Fix cover section height
If the content of the cover section has greater height than the cover element,
its contents will overflow out of it (https://cl.ly/2y2r0P1P0i3W).

I changed the height property of the cover to max-height so that the element
can grow instead of having its contents overflow out of it.